refactor: remove every thing related to the paca claim
