+++ .env.example
@@ -1,6 +1,4 @@
 STARGAZER_API_KEY=
 LATTICE_API_KEY=
-ELPACA_KEY=
-ELPACA_VALUE=
 SENTRY_DNS=
 SENTRY_DNS_WEB=
