--- source/components/CardClaim/CardClaim.scss
@@ -1,155 +1,0 @@
-@import '~assets/styles/variables';
-
-.container {
-  position: relative;
-  background-color: $white;
-  display: flex;
-  flex-direction: column;
-  padding: 12px;
-  border-radius: 8px;
-  border: 1px solid #4c2fb1;
-  margin-bottom: 8px;
-
-  .headerContainer {
-    display: flex;
-    flex-direction: row;
-    justify-content: space-between;
-    align-items: center;
-
-    .titleContainer {
-      .title {
-        color: #4c2fb1;
-      }
-    }
-  }
-
-  .close {
-    cursor: pointer;
-    position: absolute;
-    height: 30px;
-    width: 30px;
-    right: 0;
-    top: 0;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    border-top-right-radius: 7px;
-  }
-
-  .imageContainer {
-    display: flex;
-    border: 1px solid #4c2fb1;
-    border-radius: 4px;
-
-    .image {
-      height: 120px;
-      width: 70px;
-      border-radius: 4px;
-      object-fit: cover;
-    }
-  }
-
-  .infoContainer {
-    margin-top: 8px;
-    margin-bottom: 16px;
-    display: flex;
-    flex-direction: row;
-    flex-grow: 1;
-
-    .info {
-      display: flex;
-      flex-grow: 1;
-      flex-direction: column;
-      margin-left: 16px;
-
-      .infoItem {
-        display: flex;
-        flex-direction: row;
-        justify-content: space-between;
-        align-items: center;
-        padding: 2px 0px;
-
-        .itemKey {
-          font-weight: 500;
-        }
-
-        .itemValue {
-          color: #4c2fb1;
-          font-weight: 500;
-        }
-
-        .greenText {
-          color: #059669;
-        }
-
-        .claimWindowItem {
-          display: flex;
-          flex-direction: row;
-          align-items: center;
-
-          .infoIconContainer {
-            cursor: pointer;
-            margin-left: 2px;
-            height: 20px;
-            width: 20px;
-            display: flex;
-            justify-content: center;
-            align-items: center;
-
-            .infoIcon {
-              height: 14px;
-              width: 14px;
-            }
-          }
-        }
-      }
-
-      .border {
-        border-bottom: 0.3px solid lightgray;
-      }
-    }
-  }
-
-  .buttonsContainer {
-    display: flex;
-    flex-direction: row;
-    justify-content: space-between;
-
-    .buttonItem {
-      display: flex;
-      flex-grow: 1;
-
-      .button {
-        width: 100%;
-        height: 32px;
-
-        .buttonTitle {
-          font-size: 12px;
-          font-weight: 600;
-        }
-      }
-
-      .loader {
-        color: white;
-      }
-
-      .claimButton {
-        margin-right: 6px;
-      }
-
-      .learnMoreButton {
-        margin-left: 6px;
-      }
-    }
-  }
-}
-
-.sheetHeader {
-  display: flex;
-  flex-direction: row;
-  align-items: center;
-
-  .iconHeader {
-    margin-right: 6px;
-  }
-}
