--- source/components/CardClaim/constants.ts
@@ -1,29 +1,0 @@
-// Strings
-export const TITLE = 'El Paca Rewards';
-export const CLAIM_WINDOW = 'Claim window';
-export const EPOCHS_LEFT = 'Epochs left:';
-export const TIME_LEFT = 'Time left (est):';
-export const CURRENT_STREAK = 'Current claim streak:';
-export const TOTAL_CLAIMED = 'Total PACA claimed:';
-export const LEARN_MORE = 'Learn more';
-export const CLAIM = 'Claim';
-export const CLAIMED = 'Claimed';
-export const PACA = 'PACA';
-export const ACTIVE = 'Active';
-export const WAITING = 'Waiting';
-export const INFO_CONTENT =
-  'Shows current and target epochs. Epochs occur about every 1 minute, so the time is an estimate and may vary.';
-export const ALREADY_CLAIMED =
-  'You’ve already claimed your $PACA today. To keep it fair, you can only claim every 24 hours from one wallet!';
-
-// Emojis
-export const PACA_EMOJI = '🦙';
-export const TIME_EMOJI = '⏳';
-export const FIRE_EMOJI = '🔥';
-export const COLD_EMOJI = '❄️';
-export const CAMERA_EMOJI = '📸';
-export const CHECK_EMOJI = '✅️';
-
-// Numbers
-export const DOTS_SIZE = 24;
-export const HIDE_SIZE = 20;
