--- source/components/CardClaim/index.ts
@@ -1,1 +1,0 @@
-export { default } from './CardClaim.container';
