+++ source/navigation/routes.ts
@@ -58,5 +58,4 @@ export default {
   [screens.settings.importPhrase]: '/settings/wallets/import/phrase',
   [screens.settings.connectedSites]: '/settings/connectedSites',
   [screens.settings.security]: '/settings/security',
-  [screens.settings.personalize]: '/settings/personalize',
 };
