+++ source/scenes/home/Home/AssetsPanel/AssetsPanel.container.tsx
@@ -6,25 +6,15 @@ import { getAccountController } from 'utils/controllersUtils';
 import { RootState } from 'state/store';
 import IVaultState, { IAssetState } from 'state/vault/types';
 import IAssetListState from 'state/assets/types';
-import { getDagAddress } from 'utils/wallet';
-import { DAG_NETWORK } from 'constants/index';
-import userSelectors from 'selectors/userSelectors';
 import AssetsPanel from './AssetsPanel';
 
 const AssetsPanelContainer: FC = () => {
   const accountController = getAccountController();
 
   const linkTo = useLinkTo();
-  const { activeWallet, activeNetwork }: IVaultState = useSelector(
-    (state: RootState) => state.vault
-  );
+  const { activeWallet }: IVaultState = useSelector((state: RootState) => state.vault);
   const assets: IAssetListState = useSelector((state: RootState) => state.assets);
-  const isElpacaHidden = useSelector(userSelectors.getElpacaHidden);
-  const currentClaimWindow = useSelector(userSelectors.getCurrentClaimWindow);
   const activeNetworkAssets = useSelector(walletSelectors.selectActiveNetworkAssets);
-  const hasDagAddress = !!getDagAddress(activeWallet);
-  const isMainnet = activeNetwork.Constellation === DAG_NETWORK.main2.id;
-  const showCard = !isElpacaHidden && hasDagAddress && isMainnet && !!currentClaimWindow;
 
   const handleSelectAsset = (asset: IAssetState) => {
     accountController.updateAccountActiveAsset(asset);
@@ -35,19 +25,13 @@ const AssetsPanelContainer: FC = () => {
     linkTo('/asset/add');
   };
 
-  const handleHideCard = () => {
-    accountController.assetsController.setElpacaHidden(true);
-  };
-
   return (
     <AssetsPanel
       activeNetworkAssets={activeNetworkAssets}
       assets={assets}
       activeWallet={activeWallet}
-      showClaimCard={showCard}
       handleSelectAsset={handleSelectAsset}
       handleAddTokens={handleAddTokens}
-      handleHideCard={handleHideCard}
     />
   );
 };
