+++ source/scenes/home/Home/AssetsPanel/AssetsPanel.scss
@@ -124,7 +124,3 @@ section.activity {
     flex-grow: 4;
   }
 }
-
-.toastMessage2 {
-  font-weight: 500;
-}
