+++ source/scenes/home/Home/AssetsPanel/AssetsPanel.tsx
@@ -2,38 +2,16 @@ import React, { FC } from 'react';
 import AssetItem from 'components/AssetItem';
 import ButtonV3, { BUTTON_SIZES_ENUM, BUTTON_TYPES_ENUM } from 'components/ButtonV3';
 import SlidersIcon from 'assets/images/svg/sliders.svg';
-import CardClaim from 'components/CardClaim';
-import { ToastPosition, ToastType, useToast } from 'context/ToastContext';
 import IAssetPanel from './types';
 import styles from './AssetsPanel.scss';
 
 const AssetsPanel: FC<IAssetPanel> = ({
   activeNetworkAssets,
-  showClaimCard,
   assets,
   activeWallet,
   handleSelectAsset,
   handleAddTokens,
-  handleHideCard,
 }) => {
-  const { showToast } = useToast();
-
-  const handleShowToast = () => {
-    showToast({
-      type: ToastType.info,
-      position: ToastPosition.bottom,
-      title: 'El Paca rewards card hidden',
-      message1: 'To unhide your rewards card go to',
-      message2: 'Settings > Personalize > El Paca Rewards',
-      message2Style: styles.toastMessage2,
-    });
-  };
-
-  const onPressHideCard = () => {
-    handleShowToast();
-    handleHideCard();
-  };
-
   const renderAssetList = () => {
     return (
       <>
@@ -57,7 +35,6 @@ const AssetsPanel: FC<IAssetPanel> = ({
   return (
     <section className={styles.activity}>
       <div className={styles.content}>
-        {showClaimCard && <CardClaim onPressHideCard={onPressHideCard} />}
         {Object.keys(activeWallet.assets).length && <>{renderAssetList()}</>}
         <div className={styles.buttonContainer}>
           <ButtonV3
