+++ source/scenes/home/Home/AssetsPanel/styles.ts
@@ -24,9 +24,6 @@ const styles = StyleSheet.create({
     fontWeight: '600',
     color: '#5B36D3',
   },
-  toastMessage2: {
-    fontWeight: '500',
-  },
 });
 
 export default styles;
