+++ source/scenes/home/Home/AssetsPanel/types.ts
@@ -3,10 +3,8 @@ import IAssetListState from 'state/assets/types';
 
 export default interface IAssetPanel {
   activeNetworkAssets: IAssetState[];
-  showClaimCard: boolean;
   assets: IAssetListState;
   activeWallet: IWalletState;
   handleSelectAsset: (asset: IAssetState) => void;
   handleAddTokens: () => void;
-  handleHideCard: () => void;
 }
