+++ source/scenes/settings/Main/Main.scss
@@ -41,11 +41,6 @@
             width: 20px;
             height: 20px;
           }
-
-          .personalizeIconImage {
-            width: 24px;
-            height: 24px;
-          }
         }
       }
