--- source/scenes/settings/Personalize/Personalize.scss
@@ -1,26 +1,0 @@
-@import '~assets/styles/variables';
-
-.wrapper {
-  display: flex;
-  flex-direction: column;
-  width: 100%;
-  padding: 16px;
-  background-color: $gray-light-100;
-
-  .title {
-    font-weight: 600;
-    margin-bottom: 8px;
-  }
-
-  .cardContainer {
-    display: flex;
-    flex-direction: row;
-    justify-content: space-between;
-    align-items: center;
-    height: 56px;
-    border-radius: 8px;
-    border: 0.5px solid rgba(0, 0, 0, 0.12);
-    background-color: $white;
-    padding: 0px 16px;
-  }
-}
