--- source/scenes/settings/Personalize/constants.ts
@@ -1,2 +1,0 @@
-export const TITLE = 'El Paca Rewards';
-export const ITEM_TEXT = 'Hide El Paca Rewards Card';
