--- source/scenes/settings/Personalize/index.ts
@@ -1,1 +1,0 @@
-export { default } from './Personalize.container';
