--- source/scenes/settings/Personalize/types.ts
@@ -1,4 +1,0 @@
-export interface IPersonalize {
-  hidden: boolean;
-  toggleHideElpacaCard: () => void;
-}
