+++ source/scripts/Background/migration/v5_2_0.ts
@@ -1,26 +1,8 @@
 import { reload } from 'utils/browser';
 import IAssetListState from 'state/assets/types';
-import IContactBookState from 'state/contacts/types';
-import IPriceState from 'state/price/types';
-import IVaultState, { AssetSymbol, AssetType, Network } from 'state/vault/types';
-import ISwapState from 'state/swap/types';
-import IProvidersState from 'state/providers/types';
-import { INFTListState } from 'state/nfts/types';
+import { AssetSymbol, AssetType, Network } from 'state/vault/types';
 import { saveState } from 'state/localStorage';
-import { IUserState } from 'state/user/types';
 import { BASE_NETWORK, ETHEREUM_LOGO } from 'constants/index';
-
-type V5_2_0State = {
-  assets: IAssetListState;
-  nfts: INFTListState;
-  contacts: IContactBookState;
-  user: IUserState;
-  dapp: {};
-  price: IPriceState;
-  providers: IProvidersState;
-  vault: IVaultState;
-  swap: ISwapState;
-};
 
 const VERSION = '5.2.0';
 
@@ -71,7 +53,7 @@ const MigrateRunner = async (oldState: any) => {
     },
   });
   try {
-    const newState: V5_2_0State = {
+    const newState = {
       ...oldState,
       assets: {
         ...assetsUpdated,
