+++ source/utils/envUtil.ts
@@ -1,7 +1,5 @@
 export const NODE_ENV = process.env.NODE_ENV;
 export const STARGAZER_API_KEY = process.env.STARGAZER_API_KEY;
-export const ELPACA_KEY = process.env.ELPACA_KEY;
-export const ELPACA_VALUE = process.env.ELPACA_VALUE;
 export const LATTICE_API_KEY = process.env.LATTICE_API_KEY;
 export const isProd = NODE_ENV === 'production';
 export const isNative = false;
