+++ source/web/.env.sample
@@ -1,4 +1,2 @@
 STARGAZER_API_KEY=
 LATTICE_API_KEY=
-ELPACA_KEY=
-ELPACA_VALUE=
