+++ source/web/pages/App/index.tsx
@@ -18,12 +18,7 @@ global.scrypt = scryptJS.scrypt;
 
 const app = document.getElementById('app-root');
 
-if (
-  !process.env.STARGAZER_API_KEY ||
-  !process.env.LATTICE_API_KEY ||
-  !process.env.ELPACA_KEY ||
-  !process.env.ELPACA_VALUE
-) {
+if (!process.env.STARGAZER_API_KEY || !process.env.LATTICE_API_KEY) {
   throw 'Missing .env file or missing definition';
 }
